Quick start will now use Git if it is available.

+++ quick_start.sh
@@ -7,8 +7,14 @@ die () {
 
 cd ~
 [ -d .shattings ] && die Shattings already installed
-curl -L https://github.com/darconeous/shattings/tarball/master | tar xvz || die Unable to download archive
-mv darconeous-shattings* .shattings || die Unable to move archive contents
+
+if which -s git
+then
+	git clone git://github.com/darconeous/shattings.git .shattings || die Unable to clone shattings repository
+else
+	curl -L https://github.com/darconeous/shattings/tarball/master | tar xvz || die Unable to download archive
+	mv darconeous-shattings* .shattings || die Unable to move archive contents
+fi
 cd .shattings || die Unable to move into .shattings
 ./setup.sh
 . ~/.bashrc
